+++ backend/src/main/java/hanglog/community/domain/PublishedTrip.java
@@ -1,10 +1,11 @@
-package hanglog.trip.domain;
+package hanglog.community.domain;
 
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
 import hanglog.global.BaseEntity;
+import hanglog.trip.domain.Trip;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
