+++ backend/src/main/java/hanglog/community/domain/type/PublishedStatusType.java
@@ -1,4 +1,4 @@
-package hanglog.trip.domain.type;
+package hanglog.community.domain.type;
 
 public enum PublishedStatusType {
