+++ backend/src/test/java/hanglog/trip/fixture/TripFixture.java
@@ -7,8 +7,8 @@ import static hanglog.trip.fixture.DayLogFixture.LONDON_DAYLOG_1;
 import static hanglog.trip.fixture.DayLogFixture.LONDON_DAYLOG_2;
 import static hanglog.trip.fixture.DayLogFixture.LONDON_DAYLOG_EXTRA;
 
+import hanglog.community.domain.type.PublishedStatusType;
 import hanglog.trip.domain.Trip;
-import hanglog.trip.domain.type.PublishedStatusType;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
