+++ backend/src/test/java/hanglog/trip/presentation/CommunityControllerTest.java
@@ -20,10 +20,11 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import hanglog.auth.domain.MemberTokens;
 import hanglog.city.domain.City;
+import hanglog.community.dto.response.CommunitySingleTripResponse;
+import hanglog.community.dto.response.CommunityTripsResponse;
+import hanglog.community.presentation.CommunityController;
+import hanglog.community.service.CommunityService;
 import hanglog.global.ControllerTest;
-import hanglog.trip.dto.response.CommunitySingleTripResponse;
-import hanglog.trip.dto.response.CommunityTripsResponse;
-import hanglog.trip.service.CommunityService;
 import jakarta.servlet.http.Cookie;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
